apply pimpl to lightmetal_replay
